Move dialogs to dialog package
